Better handling of backslashes on JSON
Previously, in strings like: '{"a": "b\\"}', the parser would consider
the last double-quote as escaped, and thus not know when the string
ends. In fact, the first backslash escapes the second one and the
double-quote should be considered as the end of the string.

This fix makes sure the first backslash escapes the character after it
and, if the character after it is a backslash too, the second one does
not escape the next character.