Handle 'DOS' newline type in markdown format
The issue was that if a string had a DOS-type newline inside it, the
library would extract a string with a UNIX-type newline. So, when the
time came to replace the occurrences with the hashes, the handler
wouldn't find a match and the whole string would be left in the
template. When the time came to compile, the hash wouldn't be found and
so the whole process would fail.

To solve this, at the top of the `parse` method, we detect the newline
type and store it in a variable, then we convert the whole input to use
UNIX-style newlines, thus ensuring that the extracted strings will match
the ones in the input. At the end, we convert the whole template to use
the initial newline type, so that the template is consistent with the
original input.

A similar approach is applied semi-automatically when the Transcriber is
used.