Fix parsing bug with pluralized entries that include single quotes
Failing strings looked like:
"{ cnt, plural, one {This is Sam's book.} other {These are Sam's books.} }"

This would be parsed as pluralized=true, but with only the `one` plural form, which would contain a much bigger string as a translation:
"{This is Sam's book.} other {These are Sam's books.}"

Note that if there was only one single quote (e.g. inside the translation of `one`), the string was parsed successfully.