Fix testbed sample support for YML i18n
Before, the existing sample file for yml i18n did not appear on the
testbed when selecting this format; instead the textarea remained
empty. Now, it appears normally, just as it happens for other formats.

In order to fix this, some files had to be moved. This was necessary
because of the existing naming conventions that are in place
for the folder structure of the test template files.