Fix bug that did not parse all table content
A change that was introduced some time ago (#84), in order to handle
code blocks inside nested lists, turns out that it introduced another
bug related to tables. In particular, not all strings inside a table
were being be parsed.

The reason was the use of the `template` variable, which did not update
with every replacement, and thus the substring search did not yield
correct results.

Now we use the `md_template` variable, which gets updated with every
string that is removed and replaced by the replacement string, so now
the search is made upon the correct string, and now the strings inside
the table are not filtered out.

Regarding tests, there was already a table in the files used for
testing, so existing tests should have caught this. However, tests
passed fine. The reason seems to be a particular case in the the file,
regarding a javascript code block (```javascript). The "translation"
of this in the "el" language was, by convention, "el:```javascript",
and it seems that this broke the MD format, and didn't allow the
comparison to be properly made. It is not clear, at this point,
exactly why this happened. In order to mitigate this, the "faulty"
code block sample was moved to the end of the file.