Change test that messed up other tests
A couple of tests were made in a way that modified an important constant
that was never reverted to its original value. The constant is
`_RULES_ATOI`.

Previously, if a test that used this constant ran after
`HandlerTestCase`, there would be errors like:
"`RuleError: one is not a valid rule value.
Valid choices are ['I542kNfrTWnIB4wPLSnW']`".

There was no reason to change the constant in the first place,
so now these 2 tests were modified.