Add helper for plural rules placeholders
Add a helper function that replaces all serialized plural rule
placeholders (hashes), that correspond to the source language,
with those that correspond to the given target language rules.

This is useful when the template does not serialize all
pluralized content for a string (e.g. "{count, plural, <hash>}"),
but rather adds one hash placeholder for each rule
(e.g. "{count, plural, one {<hash1>} other {<hash2>}").

In the latter case, when compiling, the existing source placeholders
need to be removed completely, and the target placeholders need
to appear in their place.