Fixed po file template entries re-ordering.
When uploading a PO file in Resources, if obsoleted strings (marked with #~) are part of the file,
due to the separate handling of obsolete and non obsolete entries in polib, the template generated
at PO file handler's parsing has different order than the strings in the original file.
Thus, the information saved in the database and as a consequence in the exported file,
misses any non obsoleted strings that follow the obsoleted ones.

As a solution, we by-pass the polib.POFile 's unicode() method by not generating the template using
six.text_type(po_file) and instead provide a method in PoHandler that generates the template without
changing the order when obsoleted string are present.