Adds `EXTRACT_RAW` flag on markdown handler
The flag `EXTRACT_RAW` indicates a different behavior based on
the format. The default value of the flag is `True`, which means
that the contents of the file are extracted as is with the
escaping rules defined. In consequence if the flag is `True`
the following methods should be present: `escape`/`unescape`.

Markdown handler does not fall under this category so the
flag should be overriden to `False`.