Set `EXTRACTS_RAW` to `False` for indesign format
The flag `EXTRACTS_RAW` defaults to `True`.
When `EXTRACT_RAW` is set to `True` the handler must
also provide the `escape` and `unescaped` staticmethods.
These methods are necessary in order to convert the
parsed content (escaped aka `raw`) to unescaped (`rich`)
and back.

Some handlers does not override the default value
nor provide the necesseray methods. Since the parsed
content for those handlers is unescaped,
the `EXTRACTS_RAW` flag should be set to `False`.

Adds unit test to verify the above behavior for
all Handlers.