Unescape non-printable chars
YAML defines a set of characters that are considered to be invalid.
It accepts only printable characters: https://yaml.org/spec/1.2/spec.html#id2770814

Specifically, it says that:
On input, a YAML processor must accept all Unicode characters except those
explicitly excluded above.

The GithubMarkdown format has a YAML header, which must too conform
to these specs. Note that this is handled for the YAML part of the
GithubMarkdown format, not the YAML format.

More specifically, when parsing the YAML header, any escaped
non-printable character (i.e. a string with length > 1 in Python),
such as `\x03` and `\u0004` is converted to the equivalent character
(one-character string in Python).