Reimplementation of InDesign handler
It turns out that using UCFLib can damage the idml archive. The
indications we have of this is that files compiled by this handler:

1. Show up as "corrupted" in InDesign
2. Have messed up permissions when extracted by archive managers (tested
   on Linux and MacOS operating systems)

The new implementation operates on the archives directly. We access
'designmap.xml' to gather the story IDs, then gather all the story files
and extract the content their <Content> tags as templates. Finally we
create a new archive identical in every way possible to the old one but
with the stories replaced by the templates. A very similar process is
performed during compilation.

It turns out that trying to retain as much of the structure and content
of the resulting archive fixes our issues.

Theoretically, the output of parsing and compiling, at least in terms of
strings and individual story templates, should be the same as before,
thus the same tests should pass.