Revert DumbXml to its previous code
We had included "processing instructions" [1] in the capabilities of
DumbXml because it kind of is part of the XML specification, it appears
frequently in .idml files and if a file includes them, the previous
DumbXml couldn't handle it.

Since the InDesign handler doesn't use DumbXml after all, we revert this
change.

[1] https://en.wikipedia.org/wiki/Processing_Instruction