Escape `&` when compiling IDML (InDesign) files
"Lonely" ampersands (`&`) will make the resulting XML story fragments
invalid and subsequently InDesign will not be able to import the file
and it will appear corrupt. This commit forces the escaping of lonely
`&` to the proper XML escape sequence: `&amp;`.

If, however, a valid escape sequence appears in the string (eg `&lt;` or
`&#x0a1f;`), we will assume that the translator knew what they were
doing and we will not escape the sequence's first character.

So, for example, the following escapes will occur:

    "hello world"         -> "hello world"
    "hello &world"        -> "hello &amp;world"
    "hello &amp;world"    -> "hello &amp;world"
    "hello &lt;world"     -> "hello &lt;world"
    "hello &#x0a1f;world" -> "hello &#x0a1f;world"
    "&&#x05af;&&"         -> "&amp;&#x05af;&amp;&amp;"