Expose structured parsing result for PO files
When compiling the same file in different languages,
only the supplied stringset changes for every language
and the template remains the same.

The PoHandler for every compilation, re-parses the template
and applies the changes on top of the generated parsing result.
With this change we allow clients to reuse the parsing result
when compiling and optimize the performance by skipping the
template parsing.

Note: This change is backwards incompatible regarding the
return type of the `parse` method. The template string can be
retrieved by calling `pofile_to_str` on top of the generated
structure po.