Throw ParseError on duplicate YAML keys
Duplicate keys are not allowed, according to the YAML specification.
Under the hood, we use the PYYAML library. For the time being,
PYYAML parser accepts duplicate yaml keys and this is an issue that
has been reported by the Open Source Community
(https://github.com/yaml/pyyaml/issues/165)

We post process the generated pairs to verify that we do not have
duplicate keys.