Anchors with value are preserved in template
We fix a bug, where anchors with label and value, result in a template
string that the label is not included.

We override the default `compose_scalar_node` of YamlHandler. More
specifically, when the compose_scalar_node is called with an anchor
we have to look on the stream of the handler and split with the space
token with a max limit of 1. This way:

- The first split token is always the anchor label
- The second token is the value of anchor label. However, this second
token might also include leading spaces that should taken into
consideration and are actually part of the anchor label,
not of the anchor value.

Consider the following YAML file:

'&another_anchor        foo
   - bar
   - baz
'

Our goal is to set the start_mark.index to point to foo. Thus, we set
the start of the index to be the node.end_mark.index - the length of
`foo - bar - baz` + 7 trailing spaces.

We also fix a test (test_parse_duplicate_keys) that was randomly
crashing.