Save parent nodes positions in stringset strings
YML keys can be nested. To be able to locate parent nodes of a
string node we add in OpenString a `parents` field that contains
a list of tuples of parent YML nodes

          `(<parent_key>, <start_pos>, <end_pos>)`.

<parent_key> is the combined key (eg all parent keys concatenated
with dots to reach specific parent key) of the parent node.
<start/end_pos> is the integer position in YML content where the
value of the parent node starts/ends.

Further work should be done to store this additional information
in the DB and then use it in the compiling phase to remove for
example an empty nested key from the template and the final pro-
duced file.