Don't skip pseudo tags for first and last text elements
In faith that the tags will be reduced without affecting
the end result, the first and the last text elements
are added in the string without surrounding pseudo tag.

This creates issues when a simple two part string is parsed.
Example:
```
<bold>Hello</bold> World
```
was added without tags as
```
Hello World
``
and when compiling back the missing rule was changing the
end result to:
```
<bold>Hello World</bold>
```