Update ICU plural examples in JSON formats
Update KEYVALUEJSON and STRUCTUREDJSON examples of pluralized
ICU strings, so that the starting keyword is the same as the
variable name.

For example {count, plural, one {{cnt} item} other {{cnt} items}}
is best written as {cnt, plural, one {{cnt} item} other {{cnt} items}}.
The reason is that many ICU libraries or frameworks will probably
require the two to be identical (`cnt`), not `count` for the keyword
and `cnt` for the variable placeholder.