Fix apping values that comprise only of an anchor
The YaML parser is made to ignore anchors since they are irrelevant to
localization and they are vulnerable to "billion laughs" attacks. So,
when it encounter a:

```yaml
key: &anchor value
```

It will only consider the "value" part to be part of the extracted
string. The problem is that, when the anchor **is** encountered, the
parser _expects_ there to be a value afterwards. When this is
encountered:

```yaml
key: &anchor
```

according to the YAML spec, it should be equivalent to:

```yaml
key:
```

meaning the value of the key should be `null`.

```python
yaml.safe_load("key: &anchor")
```

However, because the handler assumes that there is a value following the
anchor, it does a `...split(' ', 1)[1]` which raises a `ValueError`.

The fix is it check against the captured YaMl node after an anchor is
encountered and, if it's a null node, return it as-is.