Better string segmentation for pptx notes
For normal slides a `p:sp` tag contains only a single `a:p` tag
so considering a paragraph to be either `p:sp` or `a:p` does not make
any difference.

For notes a `p:sp` tag can contain multiple `a:p` tags, for each new line
of the notes, so considering a paragraph to be the `a:p` tag will make
better segmented strings.

A simplified version of a notes slide is the following:

```
<p:sp>
  <a:p>
	<a:t>First line</a:t>
  </a:p>
  <a:p>
	<a:t>Second line</a:t>
  </a:p>
</p:sp>

```