Copy instead of deepcopy rPr elements
Deepcopying a beautifulsoup tag is a recursive process
that can fail due to python's `maximum recursion depth exceeded`.

In order to avoid this error instead of deep copying we just
copy the rPr tags, which doesn't copy the parent of the tag
thus preventing the exception. This doesn't cause any behavioral
change since the parent is replaced later in the process.