Fix compilation indent for block literal style
Prior to this, the indentation of scalar nodes on compilation was
determined by multiplying a given indent for the whole document
with the nesting level of a particular key. This had the effect
that at least for the block literal style nodes the indentation
in the compiled file was different from the original source file.

With the introduced fix, during compilation, we calculate the
indentation of block literal style nodes using the original
indentation in the template.