Prevent corrupted pptx compilation
When a translation contains less parts (separate by <tx> tags)
the remaining parts of the source, that are present in the source file,
must be removed from the xml.

Until now the removal was removing the text element by itself (<a:t>)
leaving other elements intact.

This resulted in the presence of some <a:r> blocks
that had not content in them and should be the reason why
the pptx files is considered corrupted.

With this commit instead of removeing the <a:t> element only we
instead remove the surround <a:r> block all together that
will lead to compilation of non corrupted files.